+++ apps/web/src/app/(app)/(homepage)/_components/newsletter.tsx
@@ -1,0 +1,5 @@
+import SignForm from '@/components/forms/sign-form';
+
+const Newsletter = () => <SignForm />;
+
+export default Newsletter;