+++ apps/web/src/app/(app)/(homepage)/_components/post-list.tsx
@@ -1,0 +1,37 @@
+import React from 'react';
+
+import Link from 'next/link';
+
+import { allPosts, type Post } from '@contentlayer';
+
+import PostCard from '@/components/blog/post-card';
+
+const PostList = () => {
+    if (allPosts.length <= 0) {
+        return (
+            <p className='text-secondary font-medium text-md leading-none'>
+                Could not retrieve any posts...
+            </p>
+        );
+    };
+
+    return (
+        <ul className='flex flex-col gap-2'>
+            {
+                allPosts
+                    .sort((a: Post, b: Post) => new Date(b.date).getTime() - new Date(a.date).getTime())
+                    .map((post: Post) => (
+                        <Link
+                            key={post._id}
+                            href={post.slug}
+                            className='border-border'
+                        >
+                            <PostCard {...post} />
+                        </Link>
+                ))
+            }
+        </ul>
+    );
+};
+
+export default PostList;