+++ apps/web/src/app/(app)/template.tsx
@@ -2,6 +2,7 @@ import { ReactNode } from "react";
 
 import Transition from "@/components/animations/transition";
 
+// Global template transition
 const Template = ({ children }: { children: ReactNode }) => (
     <Transition>{children}</Transition>
 );