+++ apps/web/src/app/(site)/(home)/page.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-import Home from '@/components/homepage';
-
-const Homepage = () => <Home />
-
-export default Homepage;