+++ apps/web/src/app/loading.tsx
@@ -2,10 +2,11 @@ import React from 'react';
 
 import Spinner from '@repo/design-system/components/spinner';
 
-export default function Loading() {
-    return (
-        <div className='flex justify-center h-screen items-center w-full'>
-            <Spinner />
-        </div>
-    );
-}
+// Routes loading fallback 
+const Loading = () => (
+    <div className='flex justify-center items-center w-full'>
+        <Spinner />
+    </div>
+);
+
+export default Loading;