+++ apps/web/src/components/homepage/newsletter.tsx
@@ -1,9 +1,0 @@
-import { Suspense } from 'react';
-
-import SignForm from '@/components/forms/sign-form';
-
-export const Newsletter = () => (
-    <Suspense fallback={null}>
-        <SignForm />
-    </Suspense>
-);