+++ apps/web/src/components/layout/nav.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 
-export const Nav = () => (
+const Nav = () => (
     <header className='mb-14 sm:mb-24 flex justify-between items-center tracking-tight w-full'>
         <div className='flex flex-col items-start'>
             <a   
                 href='/'
-                className='mb-px inline-block font-medium no-underline'
+                className='mb-px inline-block font-medium no-underline hover:text-primary'
             >
                 Gabriel Fonseca
             </a>
@@ -15,3 +15,5 @@ export const Nav = () => (
         </div>
     </header>
 );
+
+export default Nav;